+++ CHANGES.md
@@ -1,5 +1,7 @@
 # Release Notes
 
+## Next
+
 ## 0.35.1 - 2023-12-28
 
 * PR #1153: allow writing spark string to BQ datetime