+++ CHANGES.md
@@ -1,6 +1,6 @@
 # Release Notes
 
-## Next
+## 0.36.0 - 2024-01-25
 
 * PR #1155: allow lazy materialization of query on load
 * PR #1163: Added config to set the BigQuery Job timeout